refactor(web): extract domains/streaming/ transport layer from domains/chat/
- Create domains/streaming/ with stream-transport.ts, event-parser.ts,
  and stream-debug.ts (+ colocated tests)
- Eliminate domains/chat/api/client.ts barrel file — all 14 consumers
  now import directly from @/generated/api/client.gen and @/lib/api-errors
- Lift DisplayAttachment to @/types/attachment-types.ts (cross-domain type)
- Extract generic diagnostic recording to utils/diagnostics.ts; chat-domain
  diagnostics.ts retains only message-specific summary functions
- Update cross-domain allowlist for chat → streaming dependency direction
- Zero logic changes — purely structural

Closes LUM-2033
