+++ apps/web/src/domains/chat/api/interactions.ts
@@ -7,12 +7,12 @@
 
 import type { ConfirmationDecision } from "@/types/event-types";
 import type { QuestionSubmission } from "@/domains/chat/api/event-types";
+import { client } from "@/generated/api/client.gen";
 import {
   assertHasResponse,
-  client,
   extractErrorMessage,
   SDK_BASE_OPTIONS,
-} from "@/domains/chat/api/client";
+} from "@/lib/api-errors";
 
 export async function getPendingInteractions(
   assistantId: string,
