+++ apps/web/src/domains/chat/api/messages.test.ts
@@ -9,7 +9,7 @@
  */
 
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
-import { client } from "@/domains/chat/api/client";
+import { client } from "@/generated/api/client.gen";
 import { getChatHistory, normalizeContentOrder, normalizeTextSegments, postChatMessage } from "@/domains/chat/api/messages";
 
 // ---------------------------------------------------------------------------
