+++ apps/web/src/domains/chat/components/chat-route-content.tsx
@@ -121,7 +121,7 @@ import { DiskPressureBanner, type DiskPressureBannerMode } from "@/domains/chat/
 import type { VoiceInputButtonHandle } from "@/domains/chat/components/voice-input-button";
 import type { Conversation } from "@/types/conversation-types";
 import { submitQuestionResponse } from "@/domains/chat/api/interactions";
-import type { ChatEventStream } from "@/domains/chat/api/stream";
+import type { ChatEventStream } from "@/domains/streaming/stream-transport";
 
 // ---------------------------------------------------------------------------
 // Types
