+++ apps/web/src/domains/chat/hooks/stream-message-updaters.ts
@@ -11,7 +11,7 @@
 
 import type { DisplayMessage } from "@/domains/chat/utils/reconcile";
 import type { Surface } from "@/domains/chat/types/types";
-import { toDisplayAttachments } from "@/domains/chat/api/event-parser";
+import { toDisplayAttachments } from "@/domains/streaming/event-parser";
 import type { AllowlistOption, DirectoryScopeOption, ScopeOption } from "@/types/interaction-ui-types";
 import type { ChatMessageToolCall } from "@/domains/chat/api/event-types";
 import type { MessageCompleteEvent } from "@vellumai/assistant-api";
