+++ apps/web/src/domains/chat/hooks/use-empty-state-greeting.ts
@@ -12,11 +12,8 @@
 
 import { useQuery } from "@tanstack/react-query";
 
-import {
-  client,
-  assertHasResponse,
-  SDK_BASE_OPTIONS,
-} from "@/domains/chat/api/client";
+import { client } from "@/generated/api/client.gen";
+import { assertHasResponse, SDK_BASE_OPTIONS } from "@/lib/api-errors";
 import { DEFAULT_EMPTY_STATE_GREETING } from "@/domains/chat/utils/empty-state-constants";
 
 const STALE_TIME_MS = 5 * 60 * 1000;
