+++ apps/web/src/domains/chat/hooks/use-event-stream-conversation-filter.test.tsx
@@ -3,7 +3,7 @@ import { cleanup, renderHook } from "@testing-library/react";
 import { useRef, type MutableRefObject } from "react";
 
 import type { AssistantEvent } from "@/types/event-types";
-import type { ChatEventStream } from "@/domains/chat/api/stream";
+import type { ChatEventStream } from "@/domains/streaming/stream-transport";
 import {
   __resetEventBusForTesting,
   useEventBusStore,
