+++ apps/web/src/domains/chat/hooks/use-event-stream-resume-reconcile.test.tsx
@@ -2,7 +2,7 @@ import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import { cleanup, renderHook } from "@testing-library/react";
 import { useRef, type MutableRefObject } from "react";
 
-import type { ChatEventStream } from "@/domains/chat/api/stream";
+import type { ChatEventStream } from "@/domains/streaming/stream-transport";
 import {
   __resetEventBusForTesting,
   useEventBusStore,
