+++ apps/web/src/domains/streaming/stream-debug.test.ts
@@ -8,7 +8,7 @@ import {
   registerSseClient,
   resetSseDebugStateForTests,
   unregisterSseClient,
-} from "@/domains/chat/api/stream-debug";
+} from "@/domains/streaming/stream-debug";
 import type { AssistantEvent } from "@/types/event-types";
 
 beforeEach(() => {
