+++ apps/web/src/domains/streaming/stream-transport.test.ts
@@ -44,16 +44,16 @@ mock.module("@sentry/browser", () => ({
 }));
 
 import {
-  getChatDiagnosticsEvents,
-} from "@/domains/chat/utils/diagnostics";
+  getDiagnosticsEvents,
+} from "@/utils/diagnostics";
 import {
   type TurnState,
   INITIAL_TURN_STATE,
   turnReducer,
   isSending,
 } from "@/domains/chat/turn-store";
-import { parseAssistantEvent } from "@/domains/chat/api/event-parser";
-import { subscribeChatEvents, type ChatStreamReconnectCause } from "@/domains/chat/api/stream";
+import { parseAssistantEvent } from "@/domains/streaming/event-parser";
+import { subscribeChatEvents, type ChatStreamReconnectCause } from "@/domains/streaming/stream-transport";
 import { useAssistantIdentityStore } from "@/stores/assistant-identity-store";
 
 describe("polling reconciliation with state machine", () => {
@@ -629,7 +629,7 @@ describe("subscribeChatEvents idle watchdog", () => {
         ),
     ) as unknown as typeof fetch;
 
-    const eventCountBefore = getChatDiagnosticsEvents().length;
+    const eventCountBefore = getDiagnosticsEvents().length;
     const breadcrumbsBefore = sentryBreadcrumbs.length;
     const captureMessagesBefore = sentryCaptureMessages.length;
 
@@ -645,7 +645,7 @@ describe("subscribeChatEvents idle watchdog", () => {
       // Comfortably past the first watchdog fire (~50ms).
       await new Promise((r) => setTimeout(r, 200));
 
-      const newEvents = getChatDiagnosticsEvents().slice(eventCountBefore);
+      const newEvents = getDiagnosticsEvents().slice(eventCountBefore);
       const fires = newEvents.filter(
         (event) => event.kind === "sse_watchdog_fired",
       );
@@ -659,7 +659,7 @@ describe("subscribeChatEvents idle watchdog", () => {
       // The first watchdog fire happens on the very first connect
       // attempt, before any reconnect has incremented the counter.
       expect(first.details.attempt).toBe(0);
-      // Centralized platform tag is injected by recordChatDiagnostic.
+      // Centralized platform tag is injected by recordDiagnostic.
       expect(first.details.platform).toBe("web");
 
       // Sentry mirrors are how fleet data answers the L2/L3 question.
@@ -730,7 +730,7 @@ describe("subscribeChatEvents idle watchdog", () => {
         ),
     ) as unknown as typeof fetch;
 
-    const eventCountBefore = getChatDiagnosticsEvents().length;
+    const eventCountBefore = getDiagnosticsEvents().length;
     const captureMessagesBefore = sentryCaptureMessages.length;
 
     const sub = subscribeChatEvents(
@@ -751,7 +751,7 @@ describe("subscribeChatEvents idle watchdog", () => {
     try {
       await new Promise((r) => setTimeout(r, 200));
 
-      const newEvents = getChatDiagnosticsEvents().slice(eventCountBefore);
+      const newEvents = getDiagnosticsEvents().slice(eventCountBefore);
       const firstFire = newEvents.find(
         (event) => event.kind === "sse_watchdog_fired",
       );
@@ -892,7 +892,7 @@ describe("subscribeChatEvents idle watchdog", () => {
         ),
     ) as unknown as typeof fetch;
 
-    const eventCountBefore = getChatDiagnosticsEvents().length;
+    const eventCountBefore = getDiagnosticsEvents().length;
 
     const sub = subscribeChatEvents(
       "asst-heartbeat",
@@ -907,7 +907,7 @@ describe("subscribeChatEvents idle watchdog", () => {
       // idleTimeoutMs = ~120ms. 250ms gives comfortable margin.
       await new Promise((r) => setTimeout(r, 250));
 
-      const newEvents = getChatDiagnosticsEvents().slice(eventCountBefore);
+      const newEvents = getDiagnosticsEvents().slice(eventCountBefore);
       const firstFire = newEvents.find(
         (event) => event.kind === "sse_watchdog_fired",
       );
@@ -1016,7 +1016,7 @@ describe("subscribeChatEvents idle watchdog", () => {
     }) as unknown as typeof fetch;
 
     const causes: ChatStreamReconnectCause[] = [];
-    const eventCountBefore = getChatDiagnosticsEvents().length;
+    const eventCountBefore = getDiagnosticsEvents().length;
 
     const sub = subscribeChatEvents(
       "asst-stale",
@@ -1049,7 +1049,7 @@ describe("subscribeChatEvents idle watchdog", () => {
       // No sse_watchdog_fired diagnostic should have been recorded
       // for this subscription — every fetch errored before its
       // watchdog deadline, so any fire is from a stale timer.
-      const newEvents = getChatDiagnosticsEvents().slice(eventCountBefore);
+      const newEvents = getDiagnosticsEvents().slice(eventCountBefore);
       const fires = newEvents.filter(
         (event) =>
           event.kind === "sse_watchdog_fired" &&
